panorama_preprocessing/last_layer/requirements.txt
numpy==1.18.5
Pillow==8.1.2
protobuf==3.15.6
scipy==1.4.1
tensorflow-gpu==2.3.0
tensorflow-hub==0.9.0
python 3.8.8